+++ src/components/Match/GetJoinData.tsx
@@ -15,8 +15,8 @@ const GetJoinData = (props: JoinDataProps) => {
   const acceptData = acceptList?.data;
   const date = new Date();
   const nowDate = dayjs(date).format('YYYY-MM-DD');
-  const joinData: JoinData = join?.data?.data;
   const queryClient = useQueryClient();
+  const joinData: JoinData = join?.data?.data;
   const postData = props?.data;
   const status: ImageType = {
     WIN: '승리',
@@ -44,7 +44,6 @@ const GetJoinData = (props: JoinDataProps) => {
       return (
         <>
           <button
-            type='button'
             className='w-[100%] h-[48px] border border-matchgi-bordergray rounded-[4px] bg-matchgi-btnblue text-[#FFFFFF] cursor-pointer mb-[36px]'
             onClick={handleStatusChange}
           >
@@ -66,12 +65,7 @@ const GetJoinData = (props: JoinDataProps) => {
     return (
       <>
         <div className='w-full h-[200px] bg-[#FCFCFC]'>
-          <p
-            className='w-full h-[34px] font-Noto font-medium leading-[24px] text-[16 px] text-[#38393C] border border-[#EDEDED]
-  border-x-0 border-t-0 pl-[20px] mb-[22px]'
-          >
-            경기 가입 신청 목록
-          </p>
+          <p className='joinHeader'>경기 가입 신청 목록</p>
           <p className='text-sm text-[#38393C] my-8 ml-3'>신청한 사람이 없습니다.</p>
         </div>
       </>
@@ -81,12 +75,7 @@ const GetJoinData = (props: JoinDataProps) => {
   if (postData.owner === 1 && postData.matchStatus === 'ONGOING') {
     return (
       <section className='flex flex-col w-full h-full bg-[#FCFCFC]'>
-        <p
-          className='w-full h-[34px] font-Noto font-medium leading-[24px] text-[16 px] text-[#38393C] border border-[#EDEDED]
-        border-x-0 border-t-0 pl-[20px] mb-[22px]'
-        >
-          경기 가입 신청 목록
-        </p>
+        <p className='joinHeader'>경기 가입 신청 목록</p>
         {joinData &&
           joinData.userList.map((value: ImageType, id: number) => (
             <>
@@ -96,11 +85,7 @@ const GetJoinData = (props: JoinDataProps) => {
               >
                 <div className='flex flex-row w-full h-[20px] items-center gap-3 ml-[16px]'>
                   <img
-                    src={
-                      value.profileImgUrl !== null
-                        ? value.profileImgUrl
-                        : 'https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_960_720.png'
-                    }
+                    src={value.profileImgUrl !== null ? value.profileImgUrl : '/assets/images/avatar.svg'}
                     className='w-[36px] h-[36px] rounded-[100%]'
                   />
                   <span>{value.nickname}</span>
@@ -146,9 +131,7 @@ const GetJoinData = (props: JoinDataProps) => {
   if (postData.owner === -1 && postData.matchStatus === 'ONGOING') {
     return (
       <section className='flex flex-col w-full h-full bg-[#FCFCFC]'>
-        <p className='w-full h-[34px] font-Noto font-medium leading-[24px] text-[16 px] text-[#38393C] border border-[#EDEDED] border-x-0 border-t-0 pl-[20px] mb-[22px]'>
-          경기 가입 신청 목록
-        </p>
+        <p className='joinHeader'>경기 가입 신청 목록</p>
         {joinData &&
           joinData.userList.map((value: ImageType, id: number) => (
             <>
@@ -158,18 +141,12 @@ const GetJoinData = (props: JoinDataProps) => {
               >
                 <div className='flex flex-row w-full h-[20px] items-center gap-3 ml-[16px]'>
                   <img
-                    src={
-                      value.profileImgUrl !== null
-                        ? value.profileImgUrl
-                        : 'https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_960_720.png'
-                    }
+                    src={value.profileImgUrl !== null ? value.profileImgUrl : '/assets/images/avatar.svg'}
                     className='w-[36px] h-[36px] rounded-[100%]'
                   />
                   <span>{value.nickname}</span>
                 </div>
-                <div className='w-[132px] h-[36px] bg-[#14308B] rounded-[4px] text-[#FFF] flex justify-center items-center mt-[36px]'>
-                  {status[value.status]}
-                </div>
+                <div className='joinStatus'>{status[value.status]}</div>
               </div>
             </>
           ))}
@@ -184,12 +161,7 @@ const GetJoinData = (props: JoinDataProps) => {
   ) {
     return (
       <section className='flex flex-col w-full h-full bg-[#FCFCFC]'>
-        <p
-          className='w-full h-[34px] font-Noto font-medium leading-[24px] text-[16 px] text-[#38393C] border border-[#EDEDED]
-        border-x-0 border-t-0 pl-[20px] mb-[22px]'
-        >
-          경기 결과 등록
-        </p>
+        <p className='joinHeader'>경기 결과 등록</p>
         {acceptData &&
           acceptData.map((value: ImageType, id: number) => (
             <>
@@ -199,19 +171,14 @@ const GetJoinData = (props: JoinDataProps) => {
               >
                 <div className='flex flex-row w-full h-[20px] items-center gap-3 ml-[16px]'>
                   <img
-                    src={
-                      value.profileImgUrl !== null
-                        ? value.profileImgUrl
-                        : 'https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_960_720.png'
-                    }
+                    src={value.profileImgUrl !== null ? value.profileImgUrl : '/assets/images/avatar.svg'}
                     className='w-[36px] h-[36px] rounded-[100%]'
                   />
                   <span>{value.nickname}</span>
                 </div>
                 <div className='flex flex-row gap-[33px]'>
                   <button
-                    className='box-border w-[82px] h-[36px] rounded-[4px] flex justify-center items-center mt-[36px]
-                    bg-[#FFF] text-[#38393c] border border-[#C5C6CA] hover:bg-[#14308B] hover:text-[#FFF]'
+                    className='joinScore'
                     onClick={async () => {
                       await apis.updateTotalStatus(value.requestId, { status: 'WIN' });
                       queryClient.invalidateQueries(['acceptlist']);
@@ -221,8 +188,7 @@ const GetJoinData = (props: JoinDataProps) => {
                     승
                   </button>
                   <button
-                    className='box-border w-[82px] h-[36px] rounded-[4px] flex justify-center items-center mt-[36px]
-                    bg-[#FFF] text-[#38393c] border border-[#C5C6CA] hover:bg-[#14308B] hover:text-[#FFF]'
+                    className='joinScore'
                     onClick={async () => {
                       await apis.updateTotalStatus(value.requestId, { status: 'LOSE' });
                       queryClient.invalidateQueries(['acceptlist']);
@@ -232,8 +198,7 @@ const GetJoinData = (props: JoinDataProps) => {
                     패
                   </button>
                   <button
-                    className='box-border w-[82px] h-[36px] rounded-[4px] flex justify-center items-center mt-[36px]
-                    bg-[#FFF] text-[#38393c] border border-[#C5C6CA] hover:bg-[#14308B] hover:text-[#FFF]'
+                    className='joinScore'
                     onClick={async () => {
                       await apis.updateTotalStatus(value.requestId, { status: 'DRAW' });
                       queryClient.invalidateQueries(['acceptlist']);
@@ -252,12 +217,7 @@ const GetJoinData = (props: JoinDataProps) => {
   if (nowDate >= postData.matchDeadline === true && postData.matchStatus === 'MATCHEND' && acceptData?.length === 0) {
     return (
       <section className='flex flex-col w-full h-full bg-[#FCFCFC]'>
-        <p
-          className='w-full h-[34px] font-Noto font-medium leading-[24px] text-[16 px] text-[#38393C] border border-[#EDEDED]
-      border-x-0 border-t-0 pl-[20px] mb-[22px]'
-        >
-          경기 결과
-        </p>
+        <p className='joinHeader'>경기 결과</p>
         {joinData &&
           joinData.userList.map((value: ImageType, id: number) => (
             <>
@@ -269,18 +229,12 @@ const GetJoinData = (props: JoinDataProps) => {
               >
                 <div className='flex flex-row w-full h-[20px] items-center gap-3 ml-[16px]'>
                   <img
-                    src={
-                      value.profileImgUrl !== null
-                        ? value.profileImgUrl
-                        : 'https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_960_720.png'
-                    }
+                    src={value.profileImgUrl !== null ? value.profileImgUrl : '/assets/images/avatar.svg'}
                     className='w-[36px] h-[36px] rounded-[100%]'
                   />
                   <span>{value.nickname}</span>
                 </div>
-                <div className='box-border w-[132px] h-[36px] bg-[#14308B] rounded-[4px] text-[#FFF] flex justify-center items-center mt-[36px]'>
-                  {status[value.status]}
-                </div>
+                <div className='joinStatus'>{status[value.status]}</div>
               </div>
             </>
           ))}
