+++ src/components/Match/GetPostList.tsx
@@ -1,17 +1,18 @@
-import React, { FC, useRef, useEffect } from 'react';
+import React, { useRef, useEffect } from 'react';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 import { apis } from '../../apis';
 import { useParams } from 'react-router-dom';
 import { PostDataProps, JoinDataProps } from '../../typings';
-import LoadingSpinner from '../Common/loadingSpinner';
 import { StaticMap } from 'react-kakao-maps-sdk';
+import { Helmet } from 'react-helmet';
+import CopyToClipboard from 'react-copy-to-clipboard';
+import dayjs from 'dayjs';
+import LoadingSpinner from '../Common/loadingSpinner';
 import GetJoinData from './GetJoinData';
+import Dday from './Dday';
 import Review from '../Review/Review';
 import ReviewDetail from '../Review/ReviewDetail';
 import HandleJoinEdit from './HandleJoinEdit';
-import { Helmet } from 'react-helmet';
-import dayjs from 'dayjs';
-import CopyToClipboard from 'react-copy-to-clipboard';
 
 const GetPostList = () => {
   const param = useParams();
@@ -43,23 +44,6 @@ const GetPostList = () => {
       title: postData?.title,
       content: postData?.content,
     },
-  };
-  const dday = () => {
-    const now = dayjs(new Date());
-    const a = dayjs(postData?.matchDeadline);
-    const c = now.diff(a, 'day');
-    const abs = Math.abs(c);
-    if (c < 1 && a.format('YYYY-MM-DD') !== now.format('YYYY-MM-DD')) {
-      return <p className='w-[5rem] h-7 text-[#38393C]'>(D-{abs + 1})</p>;
-    } else if (c < 1 && a.format('YYYY-MM-DD') === now.format('YYYY-MM-DD')) {
-      return <p className='w-[5rem] h-7 text-[#38393C]'>(D-DAY)</p>;
-    } else {
-      return <></>;
-    }
-  };
-
-  const changeData = (data: string) => {
-    return dayjs(data).format('YYYY.MM.DD.');
   };
 
   const handleMoveScroll = () => {
@@ -99,15 +83,6 @@ const GetPostList = () => {
           content={postData.imgurls.pop() === '' || null ? '/logo512.png' : postData.imgurls.pop()}
           data-react-helmet='true'
         />
-
-        <meta name='twitter:title' content={postData.title} data-react-helmet='true' />
-        <meta name='twitter:description' content={postData.content} data-react-helmet='true' />
-        <meta
-          name='twitter:image'
-          content={postData.imgurls.pop() === '' || null ? '/logo512.png' : postData.imgurls.pop()}
-          data-react-helmet='true'
-        />
-
         <link rel='canonical' href={`https://match-gi.com/match/${postId}`} />
       </Helmet>
       <section className='flex flex-col justify-center w-full h-full bg-[#FCFCFC] font-Noto'>
@@ -135,7 +110,7 @@ const GetPostList = () => {
           <div ref={matchRef}></div>
         </div>
         <div className='flex flex-row w-full h-[29px] justify-center items-center gap-[25px] font-Noto'>
-          <button className='detail-btn' onClick={handleMoveScroll} autoFocus>
+          <button className='detail-btn' onClick={handleMoveScroll} autoFocus value='matchRef'>
             경기정보
           </button>
           <button className='detail-btn' onClick={handleMoveScroll2}>
@@ -163,11 +138,13 @@ const GetPostList = () => {
             <div>
               <p className='w-full h-7 font-Noto'>{postData.subject}</p>
               <span className='flex flex-row gap-4'>
-                <p className='w-[85px] h-7 font-SD'>{changeData(postData.matchDeadline)}</p>
+                <p className='w-[85px] h-7 font-SD'>{dayjs(postData.matchDeadline).format('YYYY.MM.DD.')}</p>
                 {postData.matchStatus === 'MATCHEND' ? (
                   <p className='w-[3.5rem] h-7 text-[#9A9B9F] font-Noto'>(마감)</p>
                 ) : (
-                  <p className='w-[50px] font-SD'>{dday()}</p>
+                  <p className='w-[50px] font-SD'>
+                    <Dday {...drill} />
+                  </p>
                 )}
               </span>
             </div>
